backend/requirements.txt
# ── Backend: FastAPI + ML + PostgreSQL ──
# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Banco de Dados
sqlalchemy==2.0.36
psycopg[binary]==3.2.5
alembic==1.14.1

# Machine Learning
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.2

# Utilitários
python-dotenv==1.0.1
pydantic-settings==2.7.1
joblib==1.4.2

# IA Generativa
google-genai==0.5.0
